moves fillDataChartRecords to fix attempted dom access in web worker
